fixing compilation warning regarding type conversion and unsafe mix of types
svn path=/trunk/yarp2/; revision=9868

+++ src/yarpmanager/gybuilder/src/application_window.cpp
@@ -1672,8 +1672,8 @@ void ApplicationWindow::onDragDataReceived(const Glib::RefPtr<Gdk::DragContext>&
                             int x, int y, const Gtk::SelectionData& data, 
                             guint info, guint time)
 {
-    x+=m_ScrollView.get_hadjustment()->get_value();
-    y+=m_ScrollView.get_vadjustment()->get_value();
+    x += (int)m_ScrollView.get_hadjustment()->get_value();
+    y += (int)m_ScrollView.get_vadjustment()->get_value();
     if((data.get_length() >= 0) && (data.get_format() == MODULE))
     {
         const guchar* name = data.get_data();
